When adding a user to a group create an account for the user if needed
Trying to add a user to a group that doesn't have an account fails with
'... is not a registered user.'. With this change adding a user to a
group does not immediately fail if there is no account for the user,
but it tries to authenticate the user and if the authentication is
successful a user account is automatically created, so that the user
can be added to the group. This only works if LDAP is used as user
backend.

With this users can be added to groups that did not log in into Gerrit
before.

Change-Id: I4f3e97f17a8210afb05e765d7082760a594dbc75
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 01de7de53e5f088bcaa5e154e2ccbbac364966fe