If subject is shortened for displaying in the UI indicate this by '...'
If a commit has a very long subject line (> 80 characters) it is
shortened when it is displayed in the Gerrit Web UI (in the change list
and in the dependencies table on the change screen). Indicate to the
user that the subject was shortened by appending '...' to the shortened
subject.

Also the subject is now cropped after a whitespace if possible.

Change-Id: Idd1f8812638997f5ab06d2c693acbb4c9f3b3a04
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: cf071283474450a0771df4c0fb09a1126582301a