Also check if user can upload in order to restore
The thought behind this is that a restoration is an upload, and
the workaround for not restoring would be to make a fresh upload
(which we don't want).

The main use case is the scenario in which a branch has been
closed to further improvements and changes shouldn't be restored.

Rather than introduce a new permission for this, keep the abandon
conditions so someone can undo an accidental abandon.

Change-Id: I31fa95c1325ac932f1fc2c90abf3a5a42c1b24db

Former-commit-id: 35ced0a374aa2a5d489e0329e7e803a9990a8497