Rename ChangeRestoreEvent to ChangeRestoredEvent
Renamed the class to be consistent with the other event class
names.

Also renamed doChangeRestoreHook to doChangeRestoredHook and
fixed a couple of typos.

The names of the default hook file and the stream event remain
the same (both `change-restored`), so this change will not have
any impact on user scripts.

Change-Id: Id50475676c33b3055d4397df1a7aa6c9e44328c6

Former-commit-id: 650f502d58f183796cfe0c11f4fe8eef1e3c4e03