Fix highlighting of outdated dependencies by red background
If a change has a dependency to an outdated patch set this dependency
should be highlighted on the ChangeScreen by setting a red background
for it. This was broken by 2ff89dae856daec04417f964c62c22d990437224 [formerly cfd7abbf5b9fda79c45771c30cc8313b12df6986]
which introduced more theme color options. Setting a background color
for odd and even rows on the changeTable was overwriting the background
color for outdated entries.

Change-Id: I2aa670d02bb09a16729d9651af61a17fce8a4adc
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 3ca1214e3e3c675df10397ef5c2c4485ce383b82