Document REST endpoint to suggest project names
With 9ec4effddf3ab418c954636f104fded5443307e5 [formerly 5cd0577828c5ec798c266c9abcf0c1dc59ebcba5] the REST endpoint to
list projects got extended to accept a prefix string as part of
the URL to limit the results to those projects starting with this
prefix. This feature was not documented in the REST API documentation.

Change-Id: Ie6a4b39700866eb5bfe8821d9e189297909726f0
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 39f2a54b5211a8be1219dab600281ae70568b323