Do not log disabled plugins as loaded
The PluginLoader scans regularly for new plugins. If a plugin is found
and it wasn't active before it logs an info message saying that the
plugin was loaded. This info message was also logged if a disabled
plugin is found, although it didn't get active. This was confusing in
case of a disabled plugin.

Change-Id: Ifbd7df151abcdefa5001797b4f42a5368966058d
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 2ec0dd34ddc40f9d923e685fb52276b37bcb165e