Show dependency list when closed change is needed by an open change
Commit 1dd9b7a0244179d5719387642a29253611037839 [formerly 7cc8ef4b01b546e2d4c1f73f34346693e906307c] made the dependency
list for an open change automatically shown if this change was
needed by another open change. However if the change was closed the
dependency list was not automatically opened if there was an open
change depending on it.

This is bad for submitting a change series, where you review and submit
one change and then you want to continue reviewing the next change in
this series. By submitting the first change it gets closed, causing its
dependency list to get closed. Now to navigate to the next change in
the series first the dependency list has to be opened manually again.

Change-Id: Ie8f222bd6c3143a71eb702b24387c19c9f282f85
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: bd806d2b90315a64f84992f269a11f53eeb1af13