Do not show "Session cookie not available" on sign in
When LDAP is used for authentication, clicking on the 'Sign In' link
opens a user/password dialog. In this dialog the "Session cookie not
available." message was always shown as warning. This warning was
pretty useless since the user was about to sign in because he had no
current session. Even worse this warning was confusing people because
they tought something went wrong.

This problem was also discussed on the Gerrit mailing list [1].

[1] https://groups.google.com/forum/#!topic/repo-discuss/j-t77m8-7I0/discussion

Change-Id: I3c92f05a22f50962f62a1fc407fad36ed0221284
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 852b395167245e3c20692363bee5e40da5cb47e2