Fix wrong 'change ... closed' message on direct push
Pushing a commit directly into the central repository with bypassing
code review wrongly results in a 'change ... closed' message if the
commit was already pushed for review and if a Change-Id is included in
the commit message. Despite of the error message the push succeeded and
the corresponding change got closed.

The wrong error message was caused by trying to close the change twice.

Bug: issue 1493
Change-Id: I979b76962e3ac9b4ea1e4c1b3b91bf7b671daacd
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 1793dcdbe6c4925c1df0d0a95b07067355c4cbee