Improve performance of ReceiveCommits by reducing RevWalk load
JGit RevWalk does not perform well when a large number of objects are
added to the start set by markStart or markUninteresting. Avoid
putting existing refs/changes/ or refs/tags/ into the RevWalk and
instead use only the refs/heads namespace and the name of the branch
used in the refs/for/ push line.

Catch existing changes by looking for their exact commit SHA-1, rather
than complete ancestory. This should have roughly the same outcome for
anyone pushing a new commit on top of an existing open change, but
with lower computional cost at the server.

Change-Id: Ie2bb9176799528f6422292f3f889e3d28cbf5135

Former-commit-id: 2fbea6c5dadab527ef53c9168c44578283df8a94