Partially revert GroupDetail to fix Admin > Groups issues
Loading Admin > Groups on a large server is horribly slow because
someone thought it was a good idea to reuse GroupDetail in
113e1a693f0834c81a581b8976ddb703aa776585 [formerly fccf928042d372f379204fcef61038f94c5b0845]. That caused the entire
group membership database to be downloaded to the browser when
showing just the list of groups, only to have the data discarded and
reloaded after selecting a specific group to view. If a server has
7k users and 17k groups, it could take ages to show the groups list.

Strip out GroupDetail and switch back to the more lightweight
AccountGroup type when showing the groups in a list format.

Since there are only 3 SYSTEM groups using well known names, and
everything else returned is of type INTERNAL because the LDAP type
no longer exists, drop the type column from the table, it isn't
really interesting anymore.

This change does drop the Owner group name column header. Displaying
it can really slow down on large servers when there are a lot of
groups. So skip displaying the owner name. Its more important to me
that the Admin > Groups table doesn't download the entire database
than having the owner group resolved in the list view. If someone
really cares enough, they can reattempt what 113e1a693f [formerly fccf928042] was trying
to accomplish, without downloading the entire membership database.

Change-Id: I884a5129d238b7aaaf6db8a886e7c0a49f73cf74

Former-commit-id: 7e031f38f7a517c1558a8936c5ce20b0781dcd38