Fix missing ObjectInserter.flush() in ReviewProjectAccess
Some JGit storage backends can't make objects available for reading
until after the ObjectInserter's flush() method is invoked. This
is already called by the commit() and commitAt() method path, but
was missed in the createRef() path used by proposing a change to
the access tab of a project.

Also disable the reflog here, as the modification is probably going
to a refs/changes/ namespace which does not bother with reflogs,
as refs are only created and never moved.

Change-Id: I769c1b98ab20c23eb496d644d8dcaebe5184e5c1

Former-commit-id: adbb4d2af56f5e327dfa4675320c5c27265abccb