Display links in PatchSetSelectBox only once
On navigation to the next/previous patch, the links in the
PatchSetSelectBox of the old patch are show twice until the new patch
is loaded. This is especially bad if the loading of the next/previous
patch fails with an exception so that the screen stays with the old
patch. Retrying in this situation to navigate to the next/previous
patch adds another instance of the links to the PatchSetSelectBox.

Change-Id: Ia5a4e2b12e22cf01ac76bd044977cbe024875595
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: e5ade76a81fb632edca062f570fb67b565fb96b4