NPE when strategy is cherry-pick and changeMerge.test enabled
Commit 31050f21 adds a new patch set when a change is submitted
with the cherry-pick strategy.  The uploader of the new patch
set is set to the ID of one of the change's approvers.

However this causes NPE if the change does not have any
approvers, which can be the case when the method is called
during mergeability testing.

Only attempt to create the new patch set if an approver was
found.

Bug: issue 1626
Change-Id: Icd0269b1a181bd89c5233a1cd044f1b5efbf1e22

Former-commit-id: 470cc60a682f0f4567172c01f42a05cae55ed8ff