Fix formatting of commit message in Internet Explorer
Since commit e9b34e480bbaa134f850f447a04f4fd80284e8d7 [formerly 8ab61558d956b72a1666b808d650a6a610d85d0a] we are relying on
the CSS attribute 'white-space' being set to 'pre' to get proper line
breaks when formatting the commit message. The problem with this is
that in Internet Explorer 'white-space: pre' is not always working [1]
and as result the whole commit message appears in a single line.

To fix this, we are now not relying on 'white-space: pre' anymore but
we ensure the line breaks by replacing them with 'p' and 'br' html
elements as it was done before.

[1] http://www.quirksmode.org/css/quirksmode.html

Bug: issue 1517
Change-Id: I4915e4f6c3967dc88c564edbd38db6d9ce9c88dc
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 1a6ff35cc09acfe14ac133fb0adf7e971e5aa436