Fix message if pushing tag is rejected because tagger is somebody else
Pushing a tag that has somebody else as tagger requires the 'Forge
Committer' access right. If this access right is missing Gerrit
currently rejects the push with "can not create new references". This
error message is misleading because the user may think that the 'Create
Reference' access right is missing which is actually assigned.

The same reject message is also returned on push of an annotated tag
if the 'Push Annotated Tag' access right is missing. Also in this case
the error message is not ideal.

Go back to the old more generic message which says 'prohibited by
Gerrit'. This error message is at least explained in the Gerrit
documentation so that the user can find out the reason. The new error
message is not documented.

Describe the missing case in the 'prohibited by Gerrit' error
documentation if pushing a tag fails because the tagger is somebody
else and the 'Forge Committer' access right is not assigned.

Change-Id: Ib2e20bbc5f1c7d47e93fc0b2c5fecfa86908419e
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 6a85c21fd86bae2828ad1b4ab4491fab8b4534ff