Show in ApprovalTable if a reviewer can't vote in a category
As a change owner you can invite reviewers on the ChangeScreen and they
will be added to the ApprovalTable. Then you wait for the review but it
might be that the reviewer has actually no permissions to vote on the
change. If the reviewer has no permissions to vote in a category this
is not visible.

For a user that cannot vote in a category the cell in the category
column of the ApprovalTable will be shown with a grey background.
A tooltip explains that the user cannot vote in this category.

Change-Id: I48485e26bab7f7f990bd569df803dfd14ae5ca5c
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: d42555ae016585e05ccc5488930dcba1bbfbf4e3