Support changes with different submit types in MergeOp
In future the submit type for a change should be controlled from
Prolog.

Currently the MergeOp algorithm assumes that the submit type is the
same for all changes to be merged. If the submit type is controlled
from Prolog changes going to the same destination branch can have
different submit types. To support this MergeOp now scans all changes
and determines their submit type, then clusters the changes with the
same submit type and loops through the submit strategies to execute
those clustered groups in turn.

Change-Id: Iadd633491aa07654f93b9ae105da93cd20927f3e
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: edb573adbbbbf1c8332645492f93b24a146fea8b