Enables the use of session management in Jetty.
This feature available is needed not by Gerrit itself
(as GerritAccount cookie is used) but more for other
plug-ins running inside the same Servlet Container.

E.g. GitBlit requires the use of Servlet sessions
(JSESSIONID) and this feature allows to avoid having
multiple JVMs and running the plug-in in the same
container as Gerrit.

Change-Id: I4d0af696a69eb9228c6b2e3eb81affb5ee80e05d

Former-commit-id: 7752a9cf2a38f47d093a9ae6c49ee71b85348ac7