Determine GERRIT_SITE from current working directory.
If the start/stop script does not find the GERRIT_SITE variable it
will to try to determine the path via dirname calls on the $0
variable. If the script is not started with a full path name this will
end up as '.' and we will not be able to find the gerrit.config file.

If we end up in this situation we instead try to determine the
GERRIT_SITE from the current working directory.

Also a small spelling mistake is fixed.

Change-Id: Ieefb686ba07113bad7ca07cfd0de3aa427914d26

Former-commit-id: b5ae7031cb843c733caee09fdae6734e0ee63a95