Allow users of ChangeTable to control styling of change rows
Currently it is hard-coded in the ChangeTable that outdated and
abandoned changes should be highlighted. With this change users of the
ChangeTable can now control from outside which changes should be
highlighted and how the highlighting should be done.

The ChangeTable which shows the dependencies on the ChangeScreen is
using the new ability so that outdated and abandoned changes are now
only highlighted in the 'Depends On' section, but not anymore in the
'Needed By' section.

Change-Id: If3140a3ad0a47d4c6813fc3aef06adb1c959f49b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: af4138b9657ca978138ca56b89380e94e7e75ab1