Send event to stream and execute hook when reviewer is added
When a new reviewer is added on a change, the `reviewer-added`
hook is executed, and the `reviewer-added` event is sent to
the event stream.

This will allow users like automated review bots to monitor the
event stream and take action only when they are explicitly added
as reviewer on a change.

Bug: Issue 1200
Change-Id: I26339360d9e1a5eabb8b6938309d5b411d6cd420

Former-commit-id: 2336bd8bac98f32330724a9c9c5fa4c096b275cb