Unset GREP_OPTIONS in the commit-msg hook
If GREP_OPTIONS is set, it can cause unexpected behaviour in the
commit-msg hook.  For example if it includes a setting like:

  --exclude=".git/*"

it causes a new Change-Id line to be appended to the commit message
on every amend.

Unsetting GREP_OPTIONS at the beginning of the hook script will
prevent such problems from occurring.

The GREP_OPTIONS setting in the user's environment is unaffected
by this change.

Bug: Issue 1569

Change-Id: If53566abf370cd526aa8159a6a243a56afa8d4e0

Former-commit-id: 5e18b9a2587f5723139590c956d1e226c7f8b33b