Add email footer token to identify if inline comments are present
Add a new mail template that is appended to the end of mails
related to users adding comments on changes.

The new template adds an extra token at the end of the mail:

  Gerrit-HasInlineComments: Yes

or

  Gerrit-HasInlineComments: No

which will be set depending on whether or not the reviewer has left
any inline comments.

This will allow users to filter their incoming notification emails
according to whether or not the review contains inline comments.

Bug: Issue 1531
Change-Id: Iafa5d69cc8a502c7afb1338dec681ca47c63338b

Former-commit-id: 85287c1663615f8bec5c34d912bd95bda8658bf4