Refactor MergeOp: extract utility methods into MergeUtil class
The MergeOp class is difficult to maintain and to extend. It is a huge
class that implements four different submit strategies. The code for
the different submit strategies is only partially the same and there
are special coding blocks which just apply to one of the submit
strategies. The class has many member variables to hold the current
state of the merge operation. They are read and updated in many methods
which makes it difficult to understand the pre and post conditions for
a method. Also some of the methods do more than their names suggest.

The goal of the refactoring is to have an abstraction for a submit
strategy. By having polymorphism for the submit type the coding for the
different submit strategies will be seperated. This will make it much
easier to add a new submit strategy because then only a clearly defined
interface has to be implemented.

The motivation for this refactoring is to support different submit
types for changes pushed to the same destination branch. The idea is to
control the submit type for a change from Prolog. The current MergeOp
algorithm assumes that the submit type is the same for all changes to
be merged. To support different submit types for changes going to the
same destination branch MergeOp should scan all changes and determine
their submit type, then cluster changes with the same submit type and
loop through the submit strategies to execute those clustered groups in
turn. This will be implemented by follow-up changes, but this is why
the submit strategy will be designed in such a way that it can be
executed several times, each time for a different set of submitted
changes.

Doing the complete refactoring in one step results in a large change
which is difficult to review. This is why the refactoring is done in
several smaller steps. As a first step for the refactoring this change
moves all merge related utility methods from MergeOp into a new
MergeUtil class. Having such util classes is a common pattern in Gerrit
(e.g. see ChangeUtil class). Having the utility methods in the
MergeUtil class makes them usable from different classes. In particular
this allows us in a next step to implement each submit strategy in an
own class and make use of the merge utility methods from there.

Change-Id: If4bb7c91b57677b58fc09bed5c56834dfd56db20
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 2b3bafa470e7a9c708ee55cc84d17e7f9fe45596