Refactor MergeOp: implement each submit strategy in an own class
This is the second step in the MergeOp refactoring. With this change
an abstraction for a submit strategy is introduced. There is a new
abstract base class that each submit strategy has to extend. By having
polymorphism for the submit type the coding for the different submit
strategies gets seperated. This makes it much easier to add a new
submit strategy because only a clearly defined interface has to be
implemented.

Change-Id: I5afe9fdc3e4c5a796cc2be715194effed9e34ddf
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 7e136c540af5a439f67633eb10746f17ea683210