Fix: Projects created outside of Gerrit could not be cloned.
The commit:

  21487e202e9bdcf37689195e56b632cebe7c889e [formerly cd94b4b7a381527f9bd216ffc7b7b3180075d16f]
  "Fix case check for project name so that symlinks work again"

...introduced an issue where projects created directly on FS
could not be cloned. Effectively rendering new and replicated
projects unreachable from gerrit slaves before flushing the
projects_list cache and rescanning the filesystem through ssh
command "ls-project".

A new check is introduced in this patch, allowing projects to
be read from FS if they are not found in the names-cache.

Change-Id: Iab33b4480179792c1f151e53fc8d037721f01264

Former-commit-id: 49658c1ec5b313e5c2798b488e5807b13f796d78