Add resumeSortKey to gerrit query summary section
When querying gerrit there is a limit for the number of results.
Add a resumeSortKey: <sortKey>" to the summary section at the end
of a query when the list of results has been truncated. Here is
how summary section would look like -

when there are more results

type: stats
rowCount: 10
runTimeMilliseconds: 14
resumeSortKey: 00202a4c00000003

and when there are no more results

type: stats
rowCount: 3
runTimeMilliseconds: 6

Change-Id: I030accb5b4c855095ebed896e868bdc8ccfd13af

Former-commit-id: 0da1af5955bc1f9a29678c81b447315abab62140