Create ref for patch set that is created when a change is cherry-picked
If Cherry Pick is chosen as submit type, on submit a new commit is
created by the cherry pick. For this commit a new patch set is created
which is added to the change. Using any of the download commands to
fetch this new patch set currently fails with 'Couldn't find remote
ref' because no ref for the new patch set is created.

Create a ref for the new patch set and trigger the replication for it.

Bug: issue 1548
Change-Id: I5500003e2d3c55c554798a864734965adfd2e8fa
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 38121faaf254823663897391e7bec2e9ad290ebd