File indentation is sometimes wrong in Side-By-Side
To reproduce:

1. Change file indendation in a patch and push.

2. Set "Ignore Whitespace" to "All" in the Prefs.

3. Empty the diff_intraline cache.

4. View the patch using Side-By-Side.

5. Change Ignore Whitespace to "None" and update the
page.

6. The lines with changed indentation are now fetched
from the base, leading gerrit to display the wrong
indentation on the right handed side.

Issue due to diff_intraline cache does not
account for changed whitespace settings, which will
trick SparseFileContent.apply() to grab lines from
left-side hunk when the line is not within the
range of the right sided hunk.

Bug: issue 816
Change-Id: I3b8c8a7aaf5d18696846d757722b3aaa4d04243e

Former-commit-id: 2424a8a9d595c87106db51da5867ed417eb8a2f5